Fixes GH-108: Console report should print start/end timestamps if requested.

+++ junit4-ant/src/main/java/com/carrotsearch/ant/tasks/junit4/events/aggregated/AggregatedTestResultEvent.java
@@ -67,7 +67,7 @@ public class AggregatedTestResultEvent implements AggregatedResultEvent {
   /**
    * Execution time in millis.
    */
-  public int getExecutionTime() {
+  public long getExecutionTime() {
     return executionTime;
   }
